+++ src/layout-forceatlas2/defaults.js
@@ -12,6 +12,5 @@ module.exports = {
   gravity: 1,
   slowDown: 1,
   barnesHutOptimize: false,
-  barnesHutTheta: 0.5,
-  weightedMass: false
+  barnesHutTheta: 0.5
 };