+++ src/layout-forceatlas2/index.js
@@ -44,7 +44,9 @@ function abstractSynchronousLayout(assign, graph, params) {
       'graphology-layout-forceatlas2: you should provide a positive number of iterations.'
     );
 
-  var getEdgeWeight = createEdgeWeightGetter(params.getEdgeWeight).fromEntry;
+  var getEdgeWeight = createEdgeWeightGetter(
+    'getEdgeWeight' in params ? params.getEdgeWeight : 'weight'
+  ).fromEntry;
 
   var outputReducer =
     typeof params.outputReducer === 'function' ? params.outputReducer : null;
@@ -59,11 +61,7 @@ function abstractSynchronousLayout(assign, graph, params) {
     );
 
   // Building matrices
-  var matrices = helpers.graphToByteArrays(
-    graph,
-    getEdgeWeight,
-    settings.weightedMass
-  );
+  var matrices = helpers.graphToByteArrays(graph, getEdgeWeight);
 
   var i;